Modification de la structure tache, car le champs temps doit etre un float
Ajout du code permettant de remplir le tableau de liste des taches

+++ ToolBoxWilliam.h
@@ -11,7 +11,7 @@
 typedef struct{
     int id;
     int col;
-    int temps;
+    float temps;
 }sTache;
 
 #endif //OPTIMISATION_D_UNE_LIGNE_D_ASSEMBLAGE_ING2_TG_2023_2024_3_8_TOOLBOXWILLIAM_H
